parisc: math-emu: rework warning options
With moving the warnings into pragmas, overriding them from the command
line no longer works, and this has to be done from code instead.

Move all the warnings that still show up with modern compilers, tested
on the oldest (4.8) and newest supported versions (10.1), and don't
disable the ones that are not reported with either.

Signed-off-by: Arnd Bergmann <arnd@arndb.de>