headers: stop including linux/slab.h from linux/crypto.h
linux/crypto.h is included from many other headers and it includes
linux/slab.h, resulting in a preprocessed file size of 1.5MB.

This is only needed for ARCH_KMALLOC_MINALIGN though, and we can
reduce the size a lot by just copying that definition.

This is slightly risky in case the ARCH_KMALLOC_MINALIGN definition
changes for some reason.

Signed-off-by: Arnd Bergmann <arnd@arndb.de>