ARM: syscall: always store thread_info->syscall
The system call number is used in a a couple of places, in particular
ptrace, seccomp and /proc/<pid>/syscall.

The last one apparently never worked reliably on ARM for tasks
that are not currently getting traced.

Storing the syscall number in the normal entry path makes it work,
as well as allowing us to see if the current system call is for
OABI compat mode, which is the next thing I want to hook into.

Since the thread_info->syscall field is not just the number any
more, it is now renamed to abi_syscall. In kernels that enable
both OABI and EABI, the upper bits of this field encode 0x900000
(__NR_OABI_SYSCALL_BASE) for OABI tasks, while normal EABI tasks
do not set the upper bits. This makes it possible to implement
the in_oabi_syscall() helper later.

All other users of thread_info->syscall go through the
syscall_get_nr() helper, which in turn filters out the ABI
bits.

Signed-off-by: Arnd Bergmann <arnd@arndb.de>