mm: remove compat numa syscalls
The compat implementations for mbind, get_mempolicy, set_mempolicy
and migrate_pages are just there to handle the subtly different
layout of bitmaps on 32-bit hosts.

The compat implementation however lacks some of the checks that
are present in the native one, in particular for checking that
the extra bits are all zero when user space has a larger mask
size than the kernel. Worse, those extra bits do not get cleared
when copying in or out of the kernel, which can lead to incorrect
data as well.

Unify the implementation to handle the compat bitmap layout
directly in the get_nodes() and copy_nodes_to_user() helpers.
Splitting out the get_bitmap() helper from get_nodes() also
helps readability of the native case.

Signed-off-by: Arnd Bergmann <arnd@arndb.de>