staging: media: atomisp: use private copy_in_user
The compat_ioctl handler for atomisp is built around
compat_alloc_user_space() and copy_in_user(), both of which I would like
to get rid of as an official interface.

I attempted to rewrite this compat handler, but it turned out
too complicated for something I cannot test.

As a hack, duplicate the x86 implementation of both functions inside
this driver to prevent it from breaking when the other one is
removed, and add the cleanup to the (already long) TODO list.

If all the other issues on that list get addressed, this one
can be as well.

Signed-off-by: Arnd Bergmann <arnd@arndb.de>