Kbuild: consolidate extra warnings in scripts/Makefile.extrawarn
The warnings are in two places: the top-level Makefile and
scripts/Makefile.extrawarn, and this is a bit confusing. Also,
a number of possibly useful warnings are disabled unconditionally
when they should really be enabled in one of the higher W= levels.

Move all these warning options to Makefile.extrawarn, and
do not disable the useful ones at least when W=2 is set.
These may be revisited in the future and moved to W=1 or the
default set.

Signed-off-by: Arnd Bergmann <arnd@arndb.de>