x86: bitops: avoid Wshadow warning for ffs()
There are thousands of warnings in a W=2 build about the ffs()
definition overriding the built-in:

arch/x86/include/asm/bitops.h:283:28: warning: declaration of 'ffs' shadows a built-in function [-Wshadow]

Use a macro to redirect callers to a function of a different
name to avoid that.

Alternatively, the function could just be removed, letting the
compiler provide the builtin.

Signed-off-by: Arnd Bergmann <arnd@arndb.de>