wireless: remove old ioctls
Commit 8bfb36766064 ("wireless: wext: remove ndo_do_ioctl fallback")
removed the ability to have private wireless ioctl commands implemented
in the normal ndo_do_ioctl handler, under the assumption that no remaining
drivers used that.

This turned out to be incorrect, as both the atmel and intersil drivers
do in fact implement a few such commands, along with SIOCDEVPRIVATE
commands and infrastructure that would allow handling them using the
more modern abstraction.

Since nobody noticed that these interfaces have been broken for over 3.5
years and the drivers are still usable without them (provided there are
still users), it seems better to just remove the now dead code rather
than trying to make it work again.  An alternative would be to just
remove the drivers entirely.

There are also a few staging drivers that got merged with the same
type of interface, which was either broken by the same commit, or
it never worked for the drivers that got merged later.

Signed-off-by: Arnd Bergmann <arnd@arndb.de>