platform/x86: amd-pmc: fix -Wundef build warning
The check for CONFIG_DEBUGFS is wrong and causes

drivers/platform/x86/amd-pmc.c:88:5: error: 'CONFIG_DEBUG_FS' is not defined, evaluates to 0 [-Werror,-Wundef]

There is really no point for an #if/#ifdef here since the debugfs
functions do nothing when CONFIG_DEBUGFS is disabled.

Remove the preprocessor check for better readability and for avoiding
the warning. amd_pmc_dev is now eight bytes larger in this
configuration, but presumably there are only a limited number instances
in any given system.

Fixes: 156ec4731cb2 ("platform/x86: amd-pmc: Add AMD platform support for S2Idle")
Signed-off-by: Arnd Bergmann <arnd@arndb.de>