sparc32: make SMP depend on SPARC_LEON_CAS
32-bit SPARC on SMP machines is the only architecture without
support for arch_futex_atomic_op_inuser() and
futex_atomic_cmpxchg_inatomic().

As far as I can tell, there are no users of SMP SPARC32 outside
of LEON, and this one only uses SMP support with out-of-tree
patches that add support for the SPARCv9 compara-and-swap-atomic
(casa) instruction in 32-bit mode.

Make SPARC32 SMP depend on that out-of-tree support to allow
removing the -ENOSYS fallback and detection logic from common
code.

Signed-off-by: Arnd Bergmann <arnd@arndb.de>