nvme: pci: rename pci specific functions
In preparation for splitting up the driver into a platform_driver and
a pci_driver front-end, rename the internal functions to clarify which
ones are pci-specific and which ones are not.

The functions that need to be duplicated in some form now start with
nvme_pci_ and take a pci_dev pointer argument where appropriate, while
the generic functions no longer have the 'pci' part in their names.

Signed-off-by: Arnd Bergmann <arnd@arndb.de>