[CHANGES REQUESTED] arm64: fix non-converging vmlinux link
When the size of the vmlinux file is just below 64MB, the kernel
may fail to link with lld, producing output such as

ld.lld: error: assignment to symbol init_pg_end does not converge
ld.lld: error: assignment to symbol __pecoff_data_size does not converge

Change the INIT_DIR_SIZE definition to not include init_pg_dir
to get a stable size calculation.

Arnd did the original report and analysis, but Ard figured what
to do about and wrote the changes to the code.

Link: https://github.com/ClangBuiltLinux/linux/issues/1219
Link: https://lore.kernel.org/linux-arm-kernel/20210927104851.GA3292@C02TD0UTHF1T.local/
Co-developed-by: Ard Biesheuvel <ardb@kernel.org>
Signed-off-by: Arnd Bergmann <arnd@arndb.de>
---
Ard, I had this in my randconfig tree with comment "Ard will
submit this with a proper changelog", but it seems we both forgot
about it, or maybe there was something wrong with it in the
end.

While looking for randconfig -Werror warnings in mainline I came
across it again and can confirm that this patch (or something like
it) is still needed. Let me know if you are happy with this version
or if you have a better description for it. I unfortunately forgot
the details of how this works.