[v2] ARM: ep93xx: don't use clang IAS for crunch
Randconfig builds with ep93xx fail with the clang integrated
assembler that does not understand the maverick crunch extensions:

arch/arm/mach-ep93xx/crunch-bits.S:94:2: error: invalid instruction
 cfstr64 mvdx0, [r1, #0] @ save 64b registers

It is unclear if anyone is still using support for crunch: gcc-4.8 dropped
it in 2012 when it was already too broken to be used reliabled. glibc
support existed as an external patch but was never merged upstream.
We could consider removing the last bits of the kernel support as well.

Turn off the integrated assembler for this file for now.

Signed-off-by: Arnd Bergmann <arnd@arndb.de>