[SUBMITTED 20210927] [RESEND] apparmor: avoid -Wempty-body warning
Building with 'make W=1' shows a warning for an empty macro:

security/apparmor/label.c: In function '__label_update':
security/apparmor/label.c:2096:59: error: suggest braces around empty body in an 'else' statement [-Werror=empty-body]
 2096 |                 AA_BUG(labels_ns(label) != labels_ns(new));

Change the macro defintion to use no_printk(), which improves
format string checking and avoids the warning.

Acked-by: John Johansen <john.johansen@canonical.com>
Link: https://lore.kernel.org/all/4e3e409e-c72e-edd5-379a-60883f166405@canonical.com/
Signed-off-by: Arnd Bergmann <arnd@arndb.de>
---
I sentn this back in April, and John replied that it's in the apparmor
tree, but the fix is still missing as of v5.15-rc2.

Please double-check and re-apply if necessary.