[SUBMITTED 20210422] ima: require CONFIG_MODULES for IMA_APPRAISE_MODSIG
When module support is disabled, CONFIG_MODULE_SIG_HASH cannot
be set either, so the build process fails with an error

certs/Makefile:52: *** Could not determine digest type to use from kernel config.  Stop.

when IMA_APPRAISE_MODSIG is set but modules are not.

Add a Kconfig dependency to prevent this configuration.

Fixes: 0165f4ca223b ("ima: enable signing of modules with build time generated key")
Signed-off-by: Arnd Bergmann <arnd@arndb.de>