m68k: mvme16x: turn #warning into a runtime warning
The unconditional #warning breaks compile-testing defconfig with
gcc -Werror:

Nobody has really cared about this one since the warning got added
in 2002, so it's unlikely that the build-time check is a significant
incentive for users to check this.

Degrade it into a runtime warning, maybe someone use these will
see that.

Fixes: d9ce1009a612 ("[PATCH] m68k: remove hwclk_time/gettod [9/20]")
Signed-off-by: Arnd Bergmann <arnd@arndb.de>