alpha: remove __attribute__((common)) from saved_config
When build testing the defconfig, I came across a gcc warning

WARNING: modpost: "saved_config" [vmlinux] is COMMON symbol

This seems to originate from an attempt to save a few bytes worth
of kernel .bss segment in linux-2.5 but isn't really necessary beyond
that.

Avoid the use of the common section by just making each instance
of this variable a separate static definition.

Fixes: 6278ba24daed ("Save and restore CIA window configuration data.")
Signed-off-by: Arnd Bergmann <arnd@arndb.de>