[SUBMITTED 20210514] Kbuild: shut up uboot mkimage output when building quietly
When building with 'make -s', most architectures produce no output
at all unless there are warnings. However, on at leat mips and nios2
there is output from /usr/bin/mkimage when that is installed:

  Image Name:   Linux-5.12.0-next-20210427-00716
  Created:      Wed Apr 28 22:03:30 2021
  Image Type:   NIOS II Linux Kernel Image (gzip compressed)
  Data Size:    2245876 Bytes = 2193.24 KiB = 2.14 MiB
  Load Address: d0000000
  Entry Point:  d0000000

Make these behave like the others and check for the '${quiet}'
variable to see if we should redirect the output to /dev/null.
Any errors and warnings from mkimage will still be seen as those
get sent to stderr.

Signed-off-by: Arnd Bergmann <arnd@arndb.de>