[SUBMITTED 20210514] mm/shuffle: fix section mismatch warning
clang sometimes decides not to inline shuffle_zone(), but it calls
a __meminit function. Without the extra __meminit annotation we get
this warning:

WARNING: modpost: vmlinux.o(.text+0x2a86d4): Section mismatch in reference from the function shuffle_zone() to the function .meminit.text:__shuffle_zone()
The function shuffle_zone() references
the function __meminit __shuffle_zone().
This is often because shuffle_zone lacks a __meminit
annotation or the annotation of __shuffle_zone is wrong.

shuffle_free_memory() did not show the same problem in my tests, but
it could happen in theory as well, so mark both as __meminit.

Signed-off-by: Arnd Bergmann <arnd@arndb.de>